Create pico_hypersonic_count
This code uses two RCWL-1601 hypersonic ranging sensors that detect traffic through a door. These sensors detect the distance to objects in front of them and if something passes under the threshold (130 cm currently but this can change) then the logic is set up to count up or down based on which sensor was passed first. 

Tuning to environment for changing this threshold and then the sleep time between hits is important. The sleep time is needed so that the counter does not count multiple times for one person.