Create pico_8x8_thermal_detection
This code uses the 8x8 thermal array sensor to detect the presence of a person in front of it. The code detects this presence by determining the number of "pixels" that have thermal signatures over the threshold. Currently the number of pixels over the threshold to trigger a detection is 5 of the 64 total. Cannot be currently used to count number in a room because it cannot determine direction of person as currently coded.